backend/requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0

# Security / JWT
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI
anthropic>=0.40.0
claude-agent-sdk>=0.1.0
openai>=1.50.0

# Utils
python-dotenv>=1.0.0
aiofiles>=23.0.0
PyYAML>=6.0

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0

# Document Generation
reportlab>=4.0.0
python-pptx>=0.6.21
python-docx>=1.1.0
matplotlib>=3.7.0

# Storage
minio>=7.2.0
jinja2>=3.0.0
